Add files via upload
In this project, we are going to implement the FLD method using the class version python code, provided in Linear Discriminant Analysis 1-1, for the classification of a dataset "spam.xlsx" into two classes, spam and non-spam emails. In the first step, we supposed that the threshold is 0 and we computed the error after implementation of FLD. 18.3 % as the error means that most of the data in the first class are classified correctly while in the second class is not. So, in the next step, we wrote a code called “threshold_play” in order to try the classification a few more times to adjust the threshold so that we reach a value at which we have a reasonable error in both classes. This value is a small negative number. As for the last step, we used the “threshold_loop” python code in order to find the best threshold value at which we could obtain the minimum error for both classes.